+++ db/migrate/20230726123929_create_blogs.rb
@@ -5,8 +5,8 @@ class CreateBlogs < ActiveRecord::Migration[7.0]
       t.text :body
       t.boolean :visible
       t.references :user, foreign_key: true
-      t.integer :likes_count
-      t.integer :comments_count
+      t.bigint :likes_count
+      t.bigint :comments_count
       t.timestamps
     end
   end
